requirements.txt
# absl-py==1.2.0
# astunparse==1.6.3
# cachetools==5.2.0
# certifi==2022.9.24
# charset-normalizer==2.1.1
# contourpy==1.0.5
# cycler==0.11.0
# editdistance==0.6.0
# filelock==3.8.0
# flatbuffers==22.9.24
# fonttools==4.37.4
# gast==0.4.0
# h5py==3.7.0
# idna==3.4
# importlib-metadata==5.0.0
# joblib==1.2.0
# keras==2.10.0
# Keras-Preprocessing==1.1.2
# kiwisolver==1.4.4
# libclang==14.0.6
# linecache2==1.0.0
# Markdown==3.4.1
# MarkupSafe==2.1.1
# oauthlib==3.2.1
# opt-einsum==3.3.0
# packaging==21.3
# pandas==1.5.0
# protobuf==3.19.6
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pyenchant==3.2.2
# pyparsing==3.0.9
# python-dateutil==2.8.2
# pytz==2022.4
# requests==2.28.1
# requests-file==1.5.1
# requests-oauthlib==1.3.1
# rsa==4.9
# scikit-learn==1.1.2
# scipy==1.9.2
# six==1.16.0
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.0
# tensorflow-estimator==2.10.0
# tensorflow-hub==0.12.0
# tensorflow-io-gcs-filesystem==0.27.0
# termcolor==2.0.1
# threadpoolctl==3.1.0
# tldextract==3.4.0
# tqdm==4.64.1
# traceback2==1.4.0
# typing-extensions==4.4.0
# 
# Werkzeug==2.2.2
# wrapt==1.14.1
# zipp==3.9.0
fastapi==0.85.0
fastapi-socketio==0.0.9
uvicorn==0.18.3
uvicorn[standard]
python-socketio==5.0.4
aiohttp==3.8.3
torch 
torchvision 
numpy 
opencv-python 
pillow 
onnxruntime
urllib3==1.26.12
grpcio==1.49.1